Moved everything into src folder
