--- .sequelizerc
@@ -1,7 +1,0 @@
-var path = require('path')
-
-module.exports = {
-  config: path.resolve('config', 'config.js'),
-  'migrations-path': path.resolve('migrations'),
-  'models-path': path.resolve('models'),
-}
