--- seeders/20231030201744-User.js
@@ -1,21 +1,0 @@
-'use strict'
-
-const bcrypt = require('bcrypt')
-
-/** @type {import('sequelize-cli').Migration} */
-module.exports = {
-  async up(queryInterface, Sequelize) {
-    await queryInterface.bulkInsert('Users', [
-      {
-        username: process.env.ADMIN_USERNAME,
-        password: bcrypt.hashSync(process.env.ADMIN_PASS, 8),
-        createdAt: new Date(0).toISOString(),
-        updatedAt: new Date(0).toISOString(),
-      },
-    ])
-  },
-
-  async down(queryInterface, Sequelize) {
-    await queryInterface.bulkDelete('Users', null, {})
-  },
-}
